requirements.txt
Django==3.2.1
djangorestframework==3.12.4
djangorestframework_simplejwt==5.0.0
drf_yasg==1.20.0
drf_yasg2==1.19.4
jose==1.0.0
mongoengine==0.24.2
pymongo==4.2.0
PyMySQL==1.0.2
python_jose==3.3.0
requests==2.22.0
django-rest-swagger==2.2.0
django-cors-headers==3.10.0
sqlparse==0.2.4
psycopg2==2.9.3
pika==1.2.0
pymysql==1.0.2
openpyxl==3.0.9